core, integration_tests: Address review commets #310
1. Make `camera_id` to start from 0.
2. Minor changes to multi-comp integration test.

+++ core/system.cpp
@@ -38,7 +38,7 @@ bool System::has_autopilot() const
     return _mavlink_system->has_autopilot();
 }
 
-bool System::has_camera(uint8_t camera_id) const
+bool System::has_camera(int camera_id) const
 {
     return _mavlink_system->has_camera(camera_id);
 }
